requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0

# XAI
lime>=0.2.0.1
scikit-image>=0.22.0

# Web app
streamlit>=1.34.0

# Utilities
Pillow>=10.0.0
numpy>=1.26.0
matplotlib>=3.8.0
requests>=2.31.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0